requirements.txt
langchain>=0.0.310
langchain_community>=0.0.3
openai>=0.27.0
faiss-cpu>=1.7.4
python-dotenv>=0.21.0
tiktoken>=0.3.0
